refactor: Update context imports and streamline code structure
- Refactored multiple components to use hooks from the new hooks directory instead of contexts for better organization.
- Updated imports in ContactLoadedGuard, PhoneVerificationCard, and various context files to reflect the new structure.
- Removed redundant code and streamlined context definitions to enhance maintainability and clarity across the codebase.
