+++ src/contexts/AnnouncementContext.tsx
@@ -3,31 +3,8 @@
 import { Announcement } from "@/domain/entities/announcement";
 import { useAuth } from "@/hooks/auth";
 import { useAnnouncementsService } from "@/hooks/useAnnouncementsService";
-import {
-  createContext,
-  ReactNode,
-  useCallback,
-  useEffect,
-  useState,
-} from "react";
-
-/**
- * AnnouncementState interface defines the shape of our announcement context value.
- */
-export interface AnnouncementState {
-  announcements: Announcement[];
-  isLoading: boolean;
-  error: string | null;
-  refetch: () => Promise<void>;
-}
-
-/**
- * Create the context with a default undefined value.
- * This forces consumers to use the useAnnouncements hook which performs a null check.
- */
-export const AnnouncementContext = createContext<AnnouncementState | undefined>(
-  undefined
-);
+import { ReactNode, useCallback, useEffect, useState } from "react";
+import { AnnouncementContext, AnnouncementState } from "./announcement.types";
 
 interface AnnouncementProviderProps {
   children: ReactNode;
