+++ src/hooks/usePhoneVerification.ts
@@ -1,4 +1,4 @@
-import { useContact } from "@/contexts/ContactContext";
+import { useContact } from "@/hooks/useContact";
 import {
   PhoneVerificationError,
   PhoneVerificationErrorType,
