+++ src/modules/dashboard/profile/pages/AttributeDetailPage.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { AttributeFormProvider } from "@/contexts/AttributeFormContext";
-import { useAttributeTypesLoading } from "@/contexts/AttributeTypeContext";
+import { useAttributeTypesLoading } from "@/hooks/useAttributeTypes";
 import { PageLoading } from "@/ui";
 import { useParams } from "react-router";
 import { AttributeForm } from "../components/AttributeForm";
